Move things from `easypy.concurrency` and `easypy.timing` to `easypy.sync`
Things that were moved and we use in Testlight:

* `TimeoutException`
* `PredicateNotSatisfied`
* `ProcessExiting`
* `break_locks`
* `TerminationSignal`
* `initialize_exception_listener`
* `raise_in_main_thread`
* `initialize_termination_listener`
* `Timebomb`
* `set_timebomb`
* `TagAlongThread`
* `SYNC`
* `LoggedRLock`
* `RWLock`
* `SoftLock`
* `skip_if_locked`
* `with_my_lock`
* `synchronized`
* `SynchronizedSingleton`
* `LoggedCondition`
* `iter_wait`
* `wait`
* `wait_progress`
* `iter_wait_progress`

Additionally, `synchronized_on_first_call` was removed and `throttled`
was moved to `easypy.timing`.
