+++ easypy/caching.py
@@ -361,36 +361,3 @@ class cached_property(object):
     def __call__(self, func):
         self._function = func
         return self
-
-
-@wrapper_decorator
-def shared_contextmanager(func):
-
-    class CtxManager():
-        def __init__(self, *args, **kwargs):
-            self.args = args
-            self.kwargs = kwargs
-            self.count = 0
-            self.func_cm = contextmanager(func)
-            self._lock = RLock()
-
-        def __enter__(self):
-            with self._lock:
-                if self.count == 0:
-                    self.ctm = self.func_cm(*self.args, **self.kwargs)
-                    self.obj = self.ctm.__enter__()
-                self.count += 1
-            return self.obj
-
-        def __exit__(self, *args):
-            with self._lock:
-                self.count -= 1
-                if self.count > 0:
-                    return
-                self.ctm.__exit__(*sys.exc_info())
-                inner.cache_pop(*self.args, **self.kwargs)
-
-    @wraps(func)
-    @locking_cache
-    def inner(*args, **kwargs):
-        return CtxManager(*args, **kwargs)
